removes 4 redux actions
