+++ packages/insomnia/src/ui/components/panes/placeholder-request-pane.tsx
@@ -1,8 +1,8 @@
 import React, { FC, useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 
+import { ForceToWorkspace } from '../../../common/import';
 import { createRequest } from '../../hooks/create-request';
-import { ForceToWorkspace } from '../../redux/modules/helpers';
 import { importFile } from '../../redux/modules/import';
 import { selectActiveWorkspace, selectSettings } from '../../redux/selectors';
 import { Hotkey } from '../hotkey';
