+++ packages/insomnia/src/ui/redux/modules/git.tsx
@@ -9,18 +9,17 @@ import * as models from '../../../models';
 import { BaseModel } from '../../../models';
 import type { GitRepository } from '../../../models/git-repository';
 import { createGitRepository } from '../../../models/helpers/git-repository-operations';
-import { isWorkspace, Workspace, WorkspaceScopeKeys } from '../../../models/workspace';
+import { isDesign, isWorkspace, Workspace, WorkspaceScope, WorkspaceScopeKeys } from '../../../models/workspace';
 import { forceWorkspaceScopeToDesign } from '../../../sync/git/force-workspace-scope-to-design';
 import { GIT_CLONE_DIR, GIT_INSOMNIA_DIR, GIT_INSOMNIA_DIR_NAME } from '../../../sync/git/git-vcs';
 import { shallowClone } from '../../../sync/git/shallow-clone';
 import { addDotGit, getOauth2FormatName, translateSSHtoHTTP } from '../../../sync/git/utils';
 import { SegmentEvent, trackSegmentEvent, vcsSegmentEventProperties } from '../../analytics';
-import { showAlert, showError, showModal } from '../../components/modals';
+import { showAlert, showError, showModal, showPrompt } from '../../components/modals';
 import { GitRepositorySettingsModal } from '../../components/modals/git-repository-settings-modal';
 import { selectActiveProject } from '../selectors';
 import { RootState } from '.';
-import { loadStart, loadStop } from './global';
-import { createWorkspace } from './workspace';
+import { activateWorkspace } from './workspace';
 
 export type UpdateGitRepositoryCallback = (arg0: { gitRepository: GitRepository }) => void;
 
@@ -49,14 +48,12 @@ export type SetupGitRepositoryCallback = (arg0: {
  * Setup a git repository against a document
  * */
 export const setupGitRepository: SetupGitRepositoryCallback = ({ createFsClient, workspace }) => {
-  return (dispatch: any) => {
+  return () => {
     trackSegmentEvent(SegmentEvent.vcsSyncStart, vcsSegmentEventProperties('git', 'setup'));
     showModal(GitRepositorySettingsModal, {
       gitRepository: null,
       onSubmitEdits: async (gitRepoPatch: Partial<GitRepository>) => {
         const providerName  = getOauth2FormatName(gitRepoPatch.credentials);
-
-        dispatch(loadStart());
 
         try {
           gitRepoPatch.needsFullClone = true;
@@ -97,8 +94,6 @@ export const setupGitRepository: SetupGitRepositoryCallback = ({ createFsClient,
           trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'setup'), providerName });
         } catch (err) {
           trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'setup', err.message), providerName });
-        } finally {
-          dispatch(loadStop());
         }
       },
     });
@@ -129,6 +124,46 @@ const createWorkspaceWithGitRepo = (gitRepo: GitRepository) => {
         },
       }),
     );
+};
+
+type OnWorkspaceCreateCallback = (arg0: Workspace) => Promise<void> | void;
+
+export const createWorkspace = ({ scope, onCreate }: {
+  scope: WorkspaceScope;
+  onCreate?: OnWorkspaceCreateCallback;
+}) => {
+  return (dispatch: any, getState: any) => {
+    const activeProject = selectActiveProject(getState());
+
+    const design = isDesign({
+      scope,
+    });
+    const title = design ? 'Design Document' : 'Request Collection';
+    const defaultValue = design ? 'my-spec.yaml' : 'My Collection';
+    const segmentEvent = design ? SegmentEvent.documentCreate : SegmentEvent.collectionCreate;
+    showPrompt({
+      title: `Create New ${title}`,
+      submitName: 'Create',
+      placeholder: defaultValue,
+      defaultValue,
+      selectText: true,
+      onComplete: async name => {
+        const flushId = await db.bufferChanges();
+        const workspace = await models.workspace.create({
+          name,
+          scope,
+          parentId: activeProject._id,
+        });
+        await models.workspace.ensureChildren(workspace);
+        await db.flushChanges(flushId);
+        if (onCreate) {
+          await onCreate(workspace);
+        }
+        await dispatch(activateWorkspace({ workspace }));
+        trackSegmentEvent(segmentEvent);
+      },
+    });
+  };
 };
 
 const cloneProblem = (message: ReactNode) => {
@@ -165,7 +200,6 @@ export const cloneGitRepository = ({ createFsClient, onComplete }: {
     showModal(GitRepositorySettingsModal, {
       gitRepository: null,
       onSubmitEdits: async (repoSettingsPatch: any) => {
-        dispatch(loadStart());
         repoSettingsPatch.needsFullClone = true;
         repoSettingsPatch.uri = translateSSHtoHTTP(repoSettingsPatch.uri);
         let fsClient = createFsClient();
@@ -182,7 +216,6 @@ export const cloneGitRepository = ({ createFsClient, onComplete }: {
               title: 'Error Cloning Repository',
               message: originalUriError.message,
             });
-            dispatch(loadStop());
             trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'clone', originalUriError.message), providerName });
             return;
           }
@@ -202,7 +235,6 @@ export const cloneGitRepository = ({ createFsClient, onComplete }: {
               title: 'Error Cloning Repository: failed to clone with and without `.git` suffix',
               message: `Failed to clone with original url (${repoSettingsPatch.uri}): ${originalUriError.message};\n\nAlso failed to clone with \`.git\` suffix added (${dotGitUri}): ${dotGitError.message}`,
             });
-            dispatch(loadStop());
             trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'clone', dotGitError.message), providerName });
             return;
           }
@@ -211,7 +243,6 @@ export const cloneGitRepository = ({ createFsClient, onComplete }: {
         // If no workspace exists, user should be prompted to create a document
         if (!(await containsInsomniaWorkspaceDir(fsClient))) {
           dispatch(noDocumentFound(repoSettingsPatch));
-          dispatch(loadStop());
           trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'clone', 'no directory found'), providerName });
           return;
         }
@@ -221,14 +252,12 @@ export const cloneGitRepository = ({ createFsClient, onComplete }: {
 
         if (workspaces.length === 0) {
           dispatch(noDocumentFound(repoSettingsPatch));
-          dispatch(loadStop());
           trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'clone', 'no workspaces found'), providerName });
           return;
         }
 
         if (workspaces.length > 1) {
           cloneProblem('Multiple workspaces found in repository; expected one.');
-          dispatch(loadStop());
           trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'clone', 'multiple workspaces found'), providerName });
           return;
         }
@@ -247,7 +276,6 @@ export const cloneGitRepository = ({ createFsClient, onComplete }: {
               before cloning.
             </>,
           );
-          dispatch(loadStop());
           trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'clone', 'workspace already exists'), providerName });
           return;
         }
@@ -290,7 +318,6 @@ export const cloneGitRepository = ({ createFsClient, onComplete }: {
 
             // Flush DB changes
             await db.flushChanges(bufferId);
-            dispatch(loadStop());
             trackSegmentEvent(SegmentEvent.vcsSyncComplete, { ...vcsSegmentEventProperties('git', 'clone'), providerName });
             onComplete?.();
           },
