+++ packages/insomnia/src/ui/redux/modules/import.ts
@@ -3,19 +3,22 @@ import { AnyAction } from 'redux';
 import { ThunkAction } from 'redux-thunk';
 
 import {
+  askToImportIntoProject,
+  askToImportIntoWorkspace,
+  askToSetWorkspaceScope,
+  ForceToWorkspace,
   importRaw,
   ImportRawConfig,
   ImportResult,
   importUri as _importUri,
 } from '../../../common/import';
 import * as models from '../../../models';
+import { DEFAULT_PROJECT_ID, Project } from '../../../models/project';
 import { Workspace, WorkspaceScope } from '../../../models/workspace';
 import { showError, showModal } from '../../components/modals';
 import { AlertModal } from '../../components/modals/alert-modal';
 import { selectActiveProject, selectProjects, selectWorkspacesWithResolvedNameForActiveProject } from '../selectors';
 import { RootState } from '.';
-import { loadStart, loadStop } from './global';
-import { askToImportIntoProject, askToImportIntoWorkspace, askToSetWorkspaceScope, ForceToWorkspace } from './helpers';
 
 export interface ImportOptions {
   workspaceId?: string;
@@ -23,6 +26,9 @@ export interface ImportOptions {
   forceToWorkspace?: ForceToWorkspace;
   forceToScope?: WorkspaceScope;
   onComplete?: () => void;
+  activeProject?: Project;
+  activeProjectWorkspaces?: Workspace[];
+  projects?: Project[];
 }
 
 const handleImportResult = (result: ImportResult, errorMessage: string) => {
@@ -48,25 +54,22 @@ const convertToRawConfig = ({
   forceToWorkspace,
   workspaceId,
   forceToProject,
-}: ImportOptions,
-state: RootState): ImportRawConfig => {
-  const activeProject = selectActiveProject(state);
-  const activeProjectWorkspaces = selectWorkspacesWithResolvedNameForActiveProject(state);
-  const projects = selectProjects(state);
+  activeProject,
+  activeProjectWorkspaces,
+  projects,
+}: ImportOptions): ImportRawConfig => {
 
   return ({
     getWorkspaceScope: askToSetWorkspaceScope(forceToScope),
     getWorkspaceId: askToImportIntoWorkspace({ workspaceId, forceToWorkspace, activeProjectWorkspaces }),
     // Currently, just return the active project instead of prompting for which project to import into
-    getProjectId: forceToProject === 'prompt' ? askToImportIntoProject({ projects, activeProject }) : () => Promise.resolve(activeProject._id),
+    getProjectId: forceToProject === 'prompt' ? askToImportIntoProject({ projects, activeProject }) : () => Promise.resolve(activeProject?._id || DEFAULT_PROJECT_ID),
   });
 };
 
 export const importFile = (
   options: ImportOptions = {},
-): ThunkAction<void, RootState, void, AnyAction> => async (dispatch, getState) => {
-  dispatch(loadStart());
-
+): ThunkAction<void, RootState, void, AnyAction> => async (_, getState) => {
   const openDialogOptions: OpenDialogOptions = {
     title: 'Import Insomnia Data',
     buttonLabel: 'Import',
@@ -94,15 +97,17 @@ export const importFile = (
 
   if (canceled) {
     // It was cancelled, so let's bail out
-    dispatch(loadStop());
     return;
   }
-
+  const state = getState();
+  const activeProject = selectActiveProject(state);
+  const activeProjectWorkspaces = selectWorkspacesWithResolvedNameForActiveProject(state);
+  const projects = selectProjects(state);
   // Let's import all the files!
   for (const filePath of filePaths) {
     try {
       const uri = `file://${filePath}`;
-      const config = convertToRawConfig(options, getState());
+      const config = convertToRawConfig({ ...options, activeProject, activeProjectWorkspaces, projects });
       const result = await _importUri(uri, config);
       handleImportResult(result, 'The file does not contain a valid specification.');
     } catch (err) {
@@ -110,8 +115,6 @@ export const importFile = (
         title: 'Import Failed',
         message: err + '',
       });
-    } finally {
-      dispatch(loadStop());
     }
   }
 
@@ -133,16 +136,18 @@ export const readFromClipBoard = () => {
 
 export const importClipBoard = (
   options: ImportOptions = {},
-): ThunkAction<void, RootState, void, AnyAction> => async (dispatch, getState) => {
-  dispatch(loadStart());
+): ThunkAction<void, RootState, void, AnyAction> => async (_, getState) => {
   const schema = readFromClipBoard();
   if (!schema) {
     return;
   }
-
+  const state = getState();
+  const activeProject = selectActiveProject(state);
+  const activeProjectWorkspaces = selectWorkspacesWithResolvedNameForActiveProject(state);
+  const projects = selectProjects(state);
   // Let's import all the paths!
   try {
-    const config = convertToRawConfig(options, getState());
+    const config = convertToRawConfig({ ...options, activeProject, activeProjectWorkspaces, projects });
     const result = await importRaw(schema, config);
     handleImportResult(result, 'Your clipboard does not contain a valid specification.');
   } catch (err) {
@@ -150,8 +155,6 @@ export const importClipBoard = (
       title: 'Import Failed',
       message: 'Your clipboard does not contain a valid specification.',
     });
-  } finally {
-    dispatch(loadStop());
   }
 
   options.onComplete?.();
@@ -160,10 +163,13 @@ export const importClipBoard = (
 export const importUri = (
   uri: string,
   options: ImportOptions = {},
-): ThunkAction<void, RootState, void, AnyAction> => async (dispatch, getState) => {
-  dispatch(loadStart());
+): ThunkAction<void, RootState, void, AnyAction> => async (_, getState) => {
+  const state = getState();
+  const activeProject = selectActiveProject(state);
+  const activeProjectWorkspaces = selectWorkspacesWithResolvedNameForActiveProject(state);
+  const projects = selectProjects(state);
   try {
-    const config = convertToRawConfig(options, getState());
+    const config = convertToRawConfig({ ...options, activeProject, activeProjectWorkspaces, projects });
     const result = await _importUri(uri, config);
     handleImportResult(result, 'The URI does not contain a valid specification.');
   } catch (err) {
@@ -171,8 +177,6 @@ export const importUri = (
       title: 'Import Failed',
       message: err + '',
     });
-  } finally {
-    dispatch(loadStop());
   }
 
   options.onComplete?.();
